src/pathway_integration.py
# Pathway Real-Time Integration for Dynamic Parking Pricing
# Template for real-time data streaming and processing

import pathway as pw
import pandas as pd
import numpy as np
from datetime import datetime
import json

class PathwayParkingPricer:
    """
    Pathway-based real-time pricing engine for parking lots
    Integrates with the dynamic pricing system for continuous updates
    """
    
    def __init__(self, base_price=10.0):
        self.base_price = base_price
        self.pricing_history = {}
        
    def setup_data_schema(self):
        """Define the data schema for incoming parking data"""
        
        # Define input schema
        parking_schema = pw.Schema.from_types(
            timestamp=pw.DatetimeNanosecond,
            lot_id=str,
            latitude=float,
            longitude=float,
            capacity=int,
            occupancy=int,
            queue_length=int,
            vehicle_type=str,
            traffic_level=float,
            is_special_day=bool
        )
        
        return parking_schema
    
    @pw.udf
    def calculate_demand_score(self, occupancy: int, capacity: int, queue_length: int, 
                              traffic_level: float, is_special_day: bool, vehicle_type: str) -> float:
        """
        UDF for calculating demand score using the demand-based model
        """
        # Normalize features
        