requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
bokeh>=2.4.0
pathway-python>=0.7.0
plotly>=5.0.0